Moving the view to the desired folders
