+++ lib/src/app.dart
@@ -21,16 +21,17 @@ import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:notify/src/models/notify_notification_quick.dart';
-import 'package:notify/src/pages/additional/edit_notification_view.dart';
-import 'package:notify/src/pages/additional/create_notification_view.dart';
+import 'package:notify/src/pages/additional/notification/edit_notification_view.dart';
+import 'package:notify/src/pages/additional/notification/create_notification_view.dart';
 import 'package:notify/src/pages/additional/list_users_view.dart';
+import 'package:notify/src/pages/additional/notification/notification_view.dart';
 import 'package:notify/src/pages/auth/auth_preview.dart';
 import 'package:notify/src/pages/auth/sign_up_view.dart';
 import 'package:notify/src/pages/brand_book_page.dart';
 import 'package:notify/src/pages/calendar/calendar_view.dart';
 import 'package:notify/src/pages/additional/color_picker_view.dart';
 import 'package:notify/src/pages/home/home_view.dart';
-import 'package:notify/src/pages/additional/edit_profile_view.dart';
+import 'package:notify/src/pages/profile/edit_profile_view.dart';
 import 'package:notify/src/pages/profile/my_profile_view.dart';
 import 'package:notify/src/pages/profile/profile_view.dart';
 import 'package:notify/src/pages/router_view.dart';
